spotifyproject/spotifyapp/urls.py
import debug_toolbar
from django.contrib.auth.views import LoginView, LogoutView
from django.urls import path, include
from .views import register_view, songs_view

urlpatterns = [
    path('', songs_view, name='songs'),
    path('register/', register_view, name='register'),
    path('login/', LoginView.as_view(template_name='login.html'), name='login'),
    path('logout/', LogoutView.as_view(next_page='/'), name='logout'),
    path("__debug__/", include("debug_toolbar.urls")),
    ]